+++ src/main/php/web/session/Sessions.class.php
@@ -12,6 +12,7 @@ abstract class Sessions {
   protected $duration= 86400;
   protected $cookie= 'session';
   protected $path= '/';
+  protected $secure= true;
 
   /**
    * Sets how long a session should last. Defaults to one day.
@@ -47,6 +48,16 @@ abstract class Sessions {
   }
 
   /**
+   * Disables to only transmit session cookie via secure connections (HTTPS).
+   *
+   * @return self
+   */
+  public function disableSecure() {
+    $this->secure= false;
+    return $this;
+  }
+
+  /**
    * Returns session duration in seconds
    *
    * @return int
@@ -68,7 +79,14 @@ abstract class Sessions {
   public function path() { return $this->path; }
 
   /**
-   * Returns session ID from request 
+   * Returns whether session cookie is set with secure flag
+   *
+   * @return  bool
+   */
+  public function isSecure() { return $this->secure; }
+
+  /**
+   * Returns session ID from request
    *
    * @param  web.Request $request
    * @return string
@@ -93,7 +111,7 @@ abstract class Sessions {
    * @return void
    */
   public function attach($id, $response) {
-    $response->cookie((new Cookie($this->cookie, $id))->maxAge($this->duration)->path($this->path));
+    $response->cookie((new Cookie($this->cookie, $id))->maxAge($this->duration)->path($this->path)->secure($this->secure));
   }
 
   /**