+++ ChangeLog.md
@@ -1,7 +1,10 @@
 Sessions for the XP Framework ChangeLog
 ========================================================================
 
-## ?.?.? / ????-??-??
+## 0.6.0 / ????-??-??
+
+* Changed session cookie to be transmitted via HTTPS only by default.
+* Added web.session.Session::disableSecure() to restore old behaviour.
 
 ## 0.5.0 / 2018-08-22